requirements.txt
astroid==1.6.5
colorama==0.3.9
Django==1.11.5
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-pagedown==1.0.5
isort==4.3.4
lazy-object-proxy==1.3.1
markdown2==2.3.5
mccabe==0.6.1
Pillow==5.2.0
psycopg2==2.7.5
pylint==1.9.2
python-http-client==3.1.0
pytz==2018.5
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.11.0
wrapt==1.10.11